[Diff][Fixed] Worktree's can leave junk branches
When kibot creates a worktree, which in some cases creates a new branch
or uses an existing branch. This branch may be old and stale, and often
not what is expected.

Instead, avoid the issue entirely, by never creating anything that could
need cleaning up, by creating a detached worktree.

Fixes #556

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>