pre_update: Honor 'board_only' (Not in schematic) footprint property.
It is quite common, to add footprints to a PCB, without adding it to the
schematic. Think Logo's or 'duplicate' connectors etc. KiCad has a flag
called 'Not in schematic' for this purpose. Lets check if this flag is
set in the PCB, and ignore any errors from this if applicable.

Fixes: #558.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>